+++ src/main/java/extra/ConPinturaMetalizada.java
@@ -2,11 +2,8 @@ package extra;
 
 import vehiculo.Vehiculo;
 
-/**
- * Created by oscar on 23/03/16.
- */
 public class ConPinturaMetalizada extends Extra {
-    public ConPinturaMetalizada(Vehiculo vehiculo, float precioExtra) {
+    public ConPinturaMetalizada(Vehiculo vehiculo, double precioExtra) {
         super(vehiculo, precioExtra);
     }
