+++ src/main/java/extra/Extra.java
@@ -2,19 +2,16 @@ package extra;
 
 import vehiculo.Vehiculo;
 
-/**
- * Created by oscar on 23/03/16.
- */
 public abstract class Extra extends Vehiculo {
     private Vehiculo vehiculo;
 
-    public Extra(Vehiculo vehiculo, float precioExtra) {
+    public Extra(Vehiculo vehiculo, double precioExtra) {
         super(precioExtra);
         this.vehiculo = vehiculo;
     }
 
     @Override
-    public float getPrecio() {
+    public double getPrecio() {
         return vehiculo.getPrecio() + super.getPrecio();
     }
